requirements.txt
anyio==4.9.0
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.0
numpy==2.2.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==22.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
